refactor the container settings model
