+++ app/models/decorators/container-configuration-decorator.ts
@@ -1,5 +1,5 @@
-import { ContainerConfiguration, ContainerRegistry } from "app/models/dtos";
 import { DecoratorBase } from "app/utils/decorators";
+import { ContainerConfiguration, ContainerRegistry } from "../container-setup";
 
 export class ContainerConfigurationDecorator extends DecoratorBase<ContainerConfiguration> {
     public type: string;
