+++ app/models/index.ts
@@ -19,6 +19,7 @@ export * from "./compute-node-error";
 export * from "./compute-node-information";
 export * from "./constraints";
 export * from "./container-lease";
+export * from "./container-setup";
 export * from "./image-reference";
 export * from "./failure-info";
 export * from "./file";
@@ -66,7 +67,6 @@ export * from "./subscription";
 export * from "./subtask-information";
 export * from "./task";
 export * from "./task-container-execution-information";
-export * from "./task-container-settings";
 export * from "./task-constraints";
 export * from "./task-dependencies";
 export * from "./task-dependency";
