+++ app/models/job-manager-task.ts
@@ -1,10 +1,10 @@
 
 import { Model, Prop, Record } from "app/core";
 import { ApplicationPackageReference } from "./application-package-reference";
+import { TaskContainerSettings, TaskContainerSettingsAttributes } from "./container-setup";
 import { NameValuePair } from "./name-value-pair";
 import { ResourceFile } from "./resource-file";
 import { TaskConstraints } from "./task-constraints";
-import { TaskContainerSettings, TaskContainerSettingsAttributes } from "./task-container-settings";
 
 export interface JobManagerTaskAttributes {
     id: string;
