+++ app/models/task.ts
@@ -5,11 +5,11 @@ import { ListProp, Model, Prop, Record } from "app/core";
 import { AffinityInformation } from "./affinity-information";
 import { ApplicationPackageReference } from "./application-package-reference";
 import { ComputeNodeInformation } from "./compute-node-information";
+import { TaskContainerSettings } from "./container-setup";
 import { MultiInstanceSettings } from "./multi-instance-settings";
 import { NameValuePair } from "./name-value-pair";
 import { ResourceFile } from "./resource-file";
 import { TaskConstraints } from "./task-constraints";
-import { TaskContainerSettings } from "./task-container-settings";
 import { TaskDependencies } from "./task-dependencies";
 import { TaskExecutionInformation } from "./task-execution-information";
 import { TaskExitConditions } from "./task-exit-conditions";
